pyproject.toml
[tool.poetry]
name = "tennis-court-finder"
version = "0.1.0"
description = "A tennis court finder application that sends alerts when court times become available"
authors = ["Andrius Jokubauskas <andrius@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
pydantic = ">=2.5.0"
httpx = ">=0.25.0"
python-multipart = ">=0.0.6"
email-validator = ">=2.1.0"
pyyaml = ">=6.0.0"
requests = ">=2.31.0"
pyjwt = ">=2.8.0"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
datamodel-code-generator = {extras = ["http"], version = ">=0.25.0"}
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"

[tool.poetry.scripts]
generate = "scripts.generate:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
